requirements.txt
Flask==3.0.0
PyGithub==2.1.1
python-dotenv==1.0.0
requests==2.31.0
packaging==23.2
APScheduler==3.10.4
pytz==2023.3.post1
pytest==7.1.1
django==3.2.0
numpy==1.19.5
pandas==1.3.0
scikit-learn==0.24.2
tensorflow==2.5.0
fastapi==0.65.2
sqlalchemy==1.4.20
pytest-cov==2.12.0
black==21.5b2
pylint==2.8.2
toml==0.10.2
xmltodict==0.13.0
pyngrok==7.0.0 